Correctly handle JSON parsing
When being deserialized, StringTable expects a mode key in the table. As the data returned from the LPix API does not contain a mode key and in addition will have a null in the err field on correct operation, we will have to actually interface with the JSONNode instead of using fromJSON and to deserialize to a StringTable as we did before.